--- src/main/java/com/kgitbank/spring/domain/model/Password.java
@@ -1,11 +1,0 @@
-package com.kgitbank.spring.domain.model;
-
-import lombok.Data;
-
-@Data
-public class Password {
-	
-	private String currentpw;
-	private String newpw;
-	
-}
