--- src/main/java/com/kgitbank/spring/domain/myprofile/controller/MyprofileService.java
@@ -1,27 +1,0 @@
-package com.kgitbank.spring.domain.myprofile.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.kgitbank.spring.domain.model.MemberVO;
-import com.kgitbank.spring.domain.myprofile.mapper.MyprofileMapper;
-
-@Service
-public class MyprofileService {
-
-	@Autowired
-	public MyprofileMapper mapper;
-	
-	public int updateMyprofile(MemberVO vo) {
-		return mapper.updateMyprofile(vo);
-	}
-	
-	public String currentpw(MemberVO vo) {
-		return mapper.currentpw(vo);
-	}
-
-	public int updatepw(MemberVO vo) {
-		return mapper.updatepw(vo);
-	}
-	
-}
