+++ src/main/java/com/kgitbank/spring/domain/myprofile/mapper/MyprofileMapper.java
@@ -8,7 +8,7 @@ public interface MyprofileMapper {
 	public int updateMyprofile(MemberVO vo);
 	
 	// SELECT pw FROM members WHERE id = #{id}
-	public String currentpw(MemberVO vo);
+	public int currentpw(MemberVO vo);
 	
 	// UPDATE members SET pw = #{pw} WHERE id = #{id}
 	public int updatepw(MemberVO vo);
