+++ src/test/java/com/kgitbank/spring/domain/myprofile/service/MyprofileServiceTest.java
@@ -9,7 +9,6 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
 import com.kgitbank.spring.domain.model.MemberVO;
-import com.kgitbank.spring.domain.myprofile.controller.MyprofileService;
 import com.kgitbank.spring.domain.test.mapper.MybatisConfigTests;
 
 import lombok.extern.log4j.Log4j;
